Turn the script a little bit more generic
Hi;

I tried to use your script as it is but it didn't work for me. I had to make some modifications and I ended up with a version which, IMHO, is more generic. For instance, I passed all directory paths to absolute path to avoid errors (specially when it pushes $startDir to `@listOfDirs). I changed the default behaviour to assume '.' as the start dir instead of '/usr/share/nginx', which I thought make more sense. I also added a little usage string to help new users. I extended the list of extensions to include HTML and JSON files, but this is optional. It should have some sort of config file to better extend the functionality with this extensions, but this is ok to me.

Nevertheless, this worked like a charm. I appreciate your work and I hope that my additions would make it even better.

Best Regards